modified some files to get ENV VAR values instead of hardcoded, moved graphics to JSON values

+++ db/schema.rb
@@ -17,7 +17,7 @@ ActiveRecord::Schema.define(version: 20151217190950) do
     t.integer  "user_id",    limit: 4
     t.string   "bank",       limit: 255
     t.string   "name",       limit: 255
-    t.float    "amount",     limit: 24
+    t.decimal    "amount",     limit: 24
     t.datetime "created_at",             null: false
     t.datetime "updated_at",             null: false
   end
